+++ CHANGES.md
@@ -21,6 +21,7 @@ Release Notes.
 * Add warning log when the event queue almost full in the access log module.
 * Reduce unessential `conntrack` query when detect new connection.
 * Reduce CPU and memory usage in the access log module.
+* Reduce handle connection event time in the access log module.
 
 #### Bug Fixes
 * Fix the base image cannot run in the arm64.